pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hmrc-tax-mcp"
version = "0.1.0"
description = "Deterministic HMRC tax rule engine exposed via Model Context Protocol"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
server = [
    "mcp[cli]>=1.0.0",
    "click>=8.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "eval-type-backport>=0.2; python_version < '3.10'",
]
extractor = [
    "anthropic>=0.25",
]

[project.scripts]
hmrc-tax-mcp = "hmrc_tax_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hmrc_tax_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true
